# docs/api/clario-dr-service.openapi.yaml
openapi: 3.1.0
info:
  title: ClarioDR Service API
  version: 1.0.0
  summary: Disaster recovery control-plane and agent-ingest contract.
  description: |
    Phase 1 contract foundation for the existing ClarioDR API. The gateway
    exposes the tenant control plane at `/api/v1/dr` and gates it with the
    `suite.datastream` entitlement. Mutating control-plane writes stage DR
    lifecycle events through `backend/internal/events/outbox` before commit.
  x-contract-phase: phase-1-foundation
  x-gateway-contract:
    id: clario-dr-service
    version: 1.0.0
    api_version: v1
    fail_closed: false
    fail_closed_profiles:
      standard: false
      regulated: true
    regulated_posture:
      env:
        - DR_DEPLOYMENT_PROFILE=regulated
        - DR_REGULATED_MODE=true
      startup_validation: fail_closed
      readiness_check: dr_runtime_config
      required_runtime:
        - Vault Transit
        - MinIO WORM recovery-point bucket
        - Redis leader election
        - Kafka event bus
        - mTLS agent ingest
        - failover driver
        - integration credential encryption key
        - command recovery driver
  x-route-sources:
    - backend/cmd/clario-dr-service/main.go
    - backend/internal/dr/handler/handler.go
    - backend/internal/dr/enroll/handler.go
    - backend/internal/dr/ingest/handler.go
    - backend/cmd/clario-dr-service/resilience.go
    - backend/cmd/clario-dr-service/intelligence.go
    - backend/cmd/clario-dr-service/orchestration.go
...
servers:
...
security:
  - bearerAuth: []
    tenantContext: []
...
paths:
...
  /groups/{groupID}/consistency-barriers:
    get:
      tags: [AppConsistency]
      operationId: listDRConsistencyBarriers
      summary: List application-consistency barrier history for a group.
      x-required-permission: dr:read
      parameters:
        - $ref: "#/components/parameters/GroupID"
      responses:
        "200":
          description: Consistency barriers.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsistencyBarrierListEnvelope"
        "404":
          $ref: "#/components/responses/NotFound"
  /groups/{groupID}/app-consistent-point:
    post:
      tags: [AppConsistency]
      operationId: createDRAppConsistentPoint
      summary: Quiesce a group and seal an application-consistent recovery point.
      x-required-permission: dr:failover
      x-outbox:
        topic: datastream.dr.events
        event_type: dr.appconsistent.barrier.created
      parameters:
        - $ref: "#/components/parameters/GroupID"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TriggerAppConsistentPointRequest"
      responses:
        "201":
          description: Consistency barrier recorded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConsistencyBarrierEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
  /recovery-points/{recoveryPointID}/cleanroom:
    get:
      tags: [CleanRoom]
      operationId: getDRCleanRoomScan
      summary: Get the latest clean-room validation verdict for a recovery point.
      x-required-permission: dr:read
      parameters:
        - $ref: "#/components/parameters/RecoveryPointID"
      responses:
        "200":
          description: Clean-room scan.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CleanRoomScanEnvelope"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /recovery-points/{recoveryPointID}/cleanroom-scan:
    post:
      tags: [CleanRoom]
      operationId: runDRCleanRoomScan
      summary: Run a clean-room validation scan for a recovery point.
      x-required-permission: dr:write
      parameters:
        - $ref: "#/components/parameters/RecoveryPointID"
      responses:
        "200":
          description: Clean-room scan result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CleanRoomScanEnvelope"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /copilot/chat:
    post:
      tags: [Copilot]
      operationId: chatDRCopilot
      summary: Submit a DR copilot chat turn.
      x-required-permission: dr:failover
      x-required-permissions: ["dr:read", "dr:failover"]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CopilotChatRequest"
      responses:
        "200":
          description: Copilot answer.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CopilotChatResultEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /copilot/sessions/{sessionID}:
    get:
      tags: [Copilot]
      operationId: getDRCopilotSession
      summary: Get a DR copilot session transcript.
      x-required-permission: dr:read
      parameters:
        - $ref: "#/components/parameters/SessionID"
      responses:
        "200":
          description: Copilot transcript.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CopilotSessionTranscriptEnvelope"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    tenantContext:
      type: apiKey
      in: header
      name: X-Tenant-ID
      description: Optional fallback when tenant context is not present in the JWT.
    agentMutualTLS:
      type: mutualTLS
  parameters:
...
    GroupID:
      name: groupID
      in: path
      required: true
      schema:
        type: string
        format: uuid
    RecoveryPointID:
      name: recoveryPointID
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    SessionID:
      name: sessionID
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    Conflict:
      description: Invalid state or already existing resource.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    ServiceUnavailable:
      description: Required dependency is not configured or unavailable.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        status:
          type: integer
        code:
          type: string
        message:
          type: string
        details: true
        request_id:
          type: string
    UUID:
      type: string
      format: uuid
    Timestamp:
      type: string
      format: date-time
...
    TriggerAppConsistentPointRequest:
      type: object
      additionalProperties: false
      properties:
        provider:
          type: string
          enum: [script, http, none]
        script:
          $ref: "#/components/schemas/ScriptQuiesceProviderRequest"
        http:
          $ref: "#/components/schemas/HTTPQuiesceProviderRequest"
        retention_until:
          $ref: "#/components/schemas/Timestamp"
    ScriptQuiesceProviderRequest:
      type: object
      required: [freeze_cmd, thaw_cmd]
      additionalProperties: false
      properties:
        freeze_cmd:
          type: string
        thaw_cmd:
          type: string
        shell:
          type: string
        timeout_sec:
          type: integer
          minimum: 0
    HTTPQuiesceProviderRequest:
      type: object
      required: [base_url]
      additionalProperties: false
      properties:
        base_url:
          type: string
          format: uri
        auth_token:
          type: string
        timeout_sec:
          type: integer
          minimum: 0
...
    ConsistencyBarrier:
      type: object
      required: [id, tenant_id, group_id, consistency_level, provider, barrier_lsn, success, quiesced, thawed, created_at]
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        tenant_id:
          $ref: "#/components/schemas/UUID"
        group_id:
          $ref: "#/components/schemas/UUID"
        recovery_point_id:
          $ref: "#/components/schemas/UUID"
        consistency_level:
          type: string
          enum: [application, crash]
        provider:
          type: string
          enum: [script, http, none]
        barrier_lsn:
          type: string
        success:
          type: boolean
        quiesced:
          type: boolean
        thawed:
          type: boolean
        detail:
          type: string
        error:
          type: string
        requested_by:
          $ref: "#/components/schemas/UUID"
        quiesce_started_at:
          $ref: "#/components/schemas/Timestamp"
        quiesce_finished_at:
          $ref: "#/components/schemas/Timestamp"
        thaw_started_at:
          $ref: "#/components/schemas/Timestamp"
        thaw_finished_at:
          $ref: "#/components/schemas/Timestamp"
        created_at:
          $ref: "#/components/schemas/Timestamp"
...
    CleanRoomChunkVerdict:
      type: object
      required: [stream_id, object_key, sandbox_path, bytes, sha256, integrity_ok, clean]
      properties:
        stream_id:
          $ref: "#/components/schemas/UUID"
        object_key:
          type: string
        sandbox_path:
          type: string
        bytes:
          type: integer
          format: int64
        sha256:
          type: string
        integrity_ok:
          type: boolean
        clean:
          type: boolean
        threat:
          type: string
    CleanRoomScan:
      type: object
      required: [id, tenant_id, recovery_point_id, group_id, verdict, scanner, chunks_scanned, bytes_scanned, detail, findings, started_at, finished_at]
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        tenant_id:
          $ref: "#/components/schemas/UUID"
        recovery_point_id:
          $ref: "#/components/schemas/UUID"
        group_id:
          $ref: "#/components/schemas/UUID"
        verdict:
          type: string
          enum: [clean, malware, integrity_failed, error]
        scanner:
          type: string
        chunks_scanned:
          type: integer
        bytes_scanned:
          type: integer
          format: int64
        detail:
          type: string
        findings:
          type: array
          items:
            $ref: "#/components/schemas/CleanRoomChunkVerdict"
        started_at:
          $ref: "#/components/schemas/Timestamp"
        finished_at:
          $ref: "#/components/schemas/Timestamp"
    CopilotChatRequest:
      type: object
      required: [message]
      additionalProperties: false
      properties:
        session_id:
          $ref: "#/components/schemas/UUID"
        message:
          type: string
          maxLength: 4000
    CopilotToolCallRecord:
      type: object
      required: [id, name, success, result_summary, latency_ms]
      properties:
        id:
          type: string
        name:
          type: string
        arguments:
          type: object
          additionalProperties: true
        success:
          type: boolean
        result_summary:
          type: string
        latency_ms:
          type: integer
    CopilotAPICall:
      type: object
      required: [method, path]
      properties:
        method:
          type: string
        path:
          type: string
        body:
          type: object
          additionalProperties: true
    CopilotProposedAction:
      type: object
      required: [kind, summary, requires_approval, api_call]
      properties:
        kind:
          type: string
        summary:
          type: string
        requires_approval:
          type: boolean
        api_call:
          $ref: "#/components/schemas/CopilotAPICall"
        approval_call:
          $ref: "#/components/schemas/CopilotAPICall"
        plan:
          type: object
          additionalProperties: true
        warnings:
          type: array
          items:
            type: string
    CopilotChatResult:
      type: object
      required: [session_id, message_id, answer, provider, model, latency_ms, iterations]
      properties:
        session_id:
          $ref: "#/components/schemas/UUID"
        message_id:
          $ref: "#/components/schemas/UUID"
        answer:
          type: string
        tool_calls:
          type: array
          items:
            $ref: "#/components/schemas/CopilotToolCallRecord"
        proposed_action:
          $ref: "#/components/schemas/CopilotProposedAction"
        provider:
          type: string
        model:
          type: string
        latency_ms:
          type: integer
        iterations:
          type: integer
    CopilotSession:
      type: object
      required: [id, tenant_id, user_id, title, provider, model, message_count, created_at, updated_at]
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        tenant_id:
          $ref: "#/components/schemas/UUID"
        user_id:
          $ref: "#/components/schemas/UUID"
        title:
          type: string
        provider:
          type: string
        model:
          type: string
        message_count:
          type: integer
        created_at:
          $ref: "#/components/schemas/Timestamp"
        updated_at:
          $ref: "#/components/schemas/Timestamp"
    CopilotMessage:
      type: object
      required: [id, session_id, tenant_id, seq, role, content, latency_ms, created_at]
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        session_id:
          $ref: "#/components/schemas/UUID"
        tenant_id:
          $ref: "#/components/schemas/UUID"
        seq:
          type: integer
        role:
          type: string
          enum: [user, assistant, tool]
        content:
          type: string
        tool_name:
          type: string
        tool_call_id:
          type: string
        tool_calls:
          type: array
          items:
            $ref: "#/components/schemas/CopilotToolCallRecord"
        proposed_action:
          $ref: "#/components/schemas/CopilotProposedAction"
        latency_ms:
          type: integer
        created_at:
          $ref: "#/components/schemas/Timestamp"
    CopilotSessionTranscript:
      type: object
      required: [session, messages]
      properties:
        session:
          $ref: "#/components/schemas/CopilotSession"
        messages:
          type: array
          items:
            $ref: "#/components/schemas/CopilotMessage"
...
    ConsistencyBarrierEnvelope:
      $ref: "#/components/schemas/DataEnvelopeConsistencyBarrier"
    ConsistencyBarrierListEnvelope:
      $ref: "#/components/schemas/DataEnvelopeConsistencyBarrierList"
...
    CleanRoomScanEnvelope:
      $ref: "#/components/schemas/DataEnvelopeCleanRoomScan"
    CopilotChatResultEnvelope:
      $ref: "#/components/schemas/DataEnvelopeCopilotChatResult"
    CopilotSessionTranscriptEnvelope:
      $ref: "#/components/schemas/DataEnvelopeCopilotSessionTranscript"
...
    DataEnvelopeConsistencyBarrier:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/ConsistencyBarrier"
    DataEnvelopeConsistencyBarrierList:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ConsistencyBarrier"
...
    DataEnvelopeCleanRoomScan:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/CleanRoomScan"
    DataEnvelopeCopilotChatResult:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/CopilotChatResult"
    DataEnvelopeCopilotSessionTranscript:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/CopilotSessionTranscript"